+++ py/surveysim/nightops.py
@@ -13,6 +13,7 @@ from desisurvey.utils import mjd2lst
 from desisurvey.nextobservation import nextFieldSelector
 from surveysim.observefield import observeField
 import desisurvey.ephemerides
+import desisurvey.config
 import desiutil.log
 
 
@@ -63,9 +64,11 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
         Updated tilesObserved table
     """
     log = desiutil.log.get_logger()
+    config = desisurvey.config.Configuration()
 
     nightOver = False
-    mjd = day_stats['MJDsunset']
+    # Start the night during bright twilight.
+    mjd = day_stats['brightdusk']
 
     if tableOutput:
         obsList = []
@@ -73,7 +76,7 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
         os.mkdir(date_string)
 
     conditions = w.getValues(mjd)
-    f = open("nightstats.dat", "a+")
+    f = open(config.get_path("nightstats.dat"), "a+")
     if conditions['OpenDome']:
         wcondsstr = "1 " + str(conditions['Seeing']) + " " + str(conditions['Transparency']) + " " + str(conditions['Clouds']) + "\n"
         f.write(wcondsstr)
@@ -90,7 +93,8 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
 
         # Initialize a moon (alt, az) interpolator using the pre-tabulated
         # ephemerides for this night.
-        moon_pos = desisurvey.ephemerides.get_moon_interpolator(day_stats)
+        moon_pos = desisurvey.ephemerides.get_object_interpolator(
+            day_stats, 'moon', altaz=True)
 
         slew = False
         ra_prev = 1.0e99
@@ -107,7 +111,7 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
                 # Compute mean to apparent to observed ra and dec???
                 airmass, tile_alt, tile_az = airMassCalculator(
                     target['RA'], target['DEC'], lst, return_altaz=True)
-                exposure = expTimeEstimator(conditions, airmass, target['Program'], target['Ebmv'], target['DESsn2'], day_stats['MoonFrac'], target['MoonDist'], target['MoonAlt'])
+                exposure = expTimeEstimator(conditions, airmass, target['Program'], target['Ebmv'], target['DESsn2'], day_stats['moon_illum_frac'], target['MoonDist'], target['MoonAlt'])
                 if exposure <= MaxExpLen:
                     status, real_exposure, real_sn2 = observeField(target, exposure)
                     real_exposure += ReadOutTime * np.floor(real_exposure/CRsplit)
@@ -123,7 +127,7 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
                         t = Time(mjd, format = 'mjd')
                         tbase = str(t.isot)
                         obsList.append((target['tileID'],  target['RA'], target['DEC'], target['PASS'], target['Program'], target['Ebmv'],
-                                       target['maxLen'], target['MoonFrac'], target['MoonDist'], target['MoonAlt'], conditions['Seeing'], conditions['Transparency'],
+                                       target['maxLen'], target['moon_illum_frac'], target['MoonDist'], target['MoonAlt'], conditions['Seeing'], conditions['Transparency'],
                                        airmass, target['DESsn2'], target['Status'],
                                        target['Exposure'], target['obsSN2'], tbase, mjd))
                     else:
@@ -136,7 +140,7 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
                         prihdr['PROGRAM '] = target['Program']
                         prihdr['EBMV    '] = target['Ebmv']
                         prihdr['MAXLEN  '] = target['maxLen']
-                        prihdr['MOONFRAC'] = target['MoonFrac']
+                        prihdr['MOONFRAC'] = target['moon_illum_frac']
                         prihdr['MOONDIST'] = target['MoonDist']
                         prihdr['MOONALT '] = target['MoonAlt']
                         prihdr['SEEING  '] = conditions['Seeing']
@@ -151,7 +155,9 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
                         nt = len(tbase)
                         prihdr['DATE-OBS'] = tbase
                         prihdr['MJD     '] = mjd
-                        filename = date_string + '/desi-exp-' + ocnt.update() + '.fits'
+                        filename = config.get_path(
+                            '{0}/desi-exp-{1}.fits'
+                            .format(date_string, ocnt.update()))
                         prihdu = pyfits.PrimaryHDU(header=prihdr)
                         prihdu.writeto(filename, clobber=True)
                 else:
@@ -163,11 +169,11 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
                 mjd += LSTres
                 slew = False
             # Check time
-            if mjd > day_stats['MJDsunrise']:
+            if mjd > day_stats['brightdawn']:
                 nightOver = True
 
     if tableOutput and len(obsList) > 0:
-        filename = 'obslist' + date_string + '.fits'
+        filename = config.get_path('obslist{0}.fits'.format(date_string))
         cols = np.rec.array(obsList,
                            names = ('TILEID  ',
                                     'RA      ',
@@ -192,12 +198,13 @@ def nightOps(day_stats, date_string, obsplan, w, ocnt, tilesObserved,
         tbhdu = pyfits.BinTableHDU.from_columns(cols)
         tbhdu.writeto(filename, clobber=True)
         # This file is to facilitate plotting
-        if os.path.exists('obslist_all.fits'):
-            obsListOld = Table.read('obslist_all.fits', format='fits')
+        all_path = config.get_path('obslist_all.fits')
+        if os.path.exists(all_path):
+            obsListOld = Table.read(all_path, format='fits')
             obsListNew = Table.read(filename, format='fits')
             obsListAll = vstack([obsListOld, obsListNew])
-            obsListAll.write('obslist_all.fits', format='fits', overwrite=True)
+            obsListAll.write(all_path, format='fits', overwrite=True)
         else:
-            copyfile(filename, 'obslist_all.fits')
+            copyfile(filename, all_path)
 
     return tilesObserved