+++ py/surveysim/scripts/surveysim.py
@@ -8,6 +8,8 @@ import os
 
 import desiutil.log
 
+import desisurvey.config
+
 import surveysim.simulator
 
 
@@ -19,10 +21,10 @@ def parse(options=None):
     parser.add_argument('-v', '--verbose', action='store_true',
         help='provide verbose output on progress')
     parser.add_argument(
-        '--start', type=str, default='2019-08-28', metavar='DATE',
+        '--start', type=str, default=None, metavar='DATE',
         help='survey starts on the evening of this day, formatted as YYYY-MM-DD')
     parser.add_argument(
-        '--stop', type=str, default='2020-07-14', metavar='DATE',
+        '--stop', type=str, default=None, metavar='DATE',
         help='survey stops on the morning of this day, formatted as YYYY-MM-DD')
     parser.add_argument(
         '--seed', type=int, default=123, metavar='N',
@@ -36,6 +38,9 @@ def parse(options=None):
     parser.add_argument(
         '--resume', default=None, metavar='FILENAME',
         help='Name of saved observations for resuming a simulation')
+    parser.add_argument(
+        '--output-path', default=None, metavar='PATH',
+        help='Output path where output files should be written')
 
     if options is None:
         args = parser.parse_args()
@@ -43,14 +48,22 @@ def parse(options=None):
         args = parser.parse_args(options)
 
     # Validate start/stop date args and covert to datetime objects.
-    try:
-        args.start = datetime.datetime.strptime(args.start, '%Y-%m-%d')
-    except ValueError as e:
-        raise ValueError('Invalid start-date: {0}'.format(e))
-    try:
-        args.stop = datetime.datetime.strptime(args.stop, '%Y-%m-%d')
-    except ValueError as e:
-        raise ValueError('Invalid stop-date: {0}'.format(e))
+    # Unspecified values are taken from our config.
+    config = desisurvey.config.Configuration()
+    if args.start is None:
+        args.start = config.first_day()
+    else:
+        try:
+            args.start = datetime.datetime.strptime(args.start, '%Y-%m-%d').date()
+        except ValueError as e:
+            raise ValueError('Invalid start-date: {0}'.format(e))
+    if args.stop is None:
+        args.stop = config.last_day()
+    else:
+        try:
+            args.stop = datetime.datetime.strptime(args.stop, '%Y-%m-%d').date()
+        except ValueError as e:
+            raise ValueError('Invalid stop-date: {0}'.format(e))
     if args.start >= args.stop:
         raise ValueError('Expected start-date < stop-date.')
 
@@ -75,14 +88,23 @@ def main(args):
         os.rename('obslist_all.fits', 'obslist_all_save.fits')
         log.info('Renamed obslist_all.fits to obslist_all_save.fits')
 
+    # Set the output path if requested.
+    config = desisurvey.config.Configuration()
+    if args.output_path is not None:
+        config.set_output_path(args.output_path)
+
+    # Create the simulator.
     simulator = surveysim.simulator.Simulator(
         args.start, args.stop, args.seed, tilesubset=None,
         use_jpl=args.use_jpl, tile_file=args.resume)
 
+    # Simulate each night until the survey is complete or the last
+    # day is reached.
     while simulator.next_day():
         pass
 
     if args.save is not None:
+        args.save = config.get_path(args.save)
         log.info('Saving observed tiles to {0}'.format(args.save))
         simulator.tilesObserved.write(
             args.save, format='fits', overwrite=True)