+++ doc/changes.rst
@@ -18,6 +18,8 @@ surveysim change log
 * Adds unit tests
 * removes data/tile-info.fits (not used here; was moved to desisurvey)
 * adds nightops.py (from desisurvey, used here but not there)
+* create surveysim command-line script
+* use new desisurvey config machinery
 
 0.3.1 (2016-12-21)
 ------------------
@@ -46,4 +48,3 @@ surveysim change log
 ---------------
 
 * No changes.rst yet
-